Fixes #246
Additional work:

 * Breakes the handler out into its own class for better SRP (fixes
#240)
 * Explicitly aligns JSON.NET serializer settings with the NetFx version
